Introduce Opcode handling and Frame decoding
A temporary solution for handling and transforming Opcodes was
introduced. The real solution will likely involve macros and/or
generated code to keep thigns in sync as the enum grows in size.

Header frames can now be decoded successfully. The implementation is not
complete, but initial unit tests show that the proof-of-concept works.