Project architecture, basic parsing and more tests
The architecture of the project lacked organization and logic was
repeated several times in various modules. As such, the modules of this
project were moved around to represent a properly organized project.
More work will be done to continue improving the architecture of the
project as this is a key aspect in creating maintainable software.

The system architecture will be improved in the next coming commits. The
other priority is proper parsing utilities. This will allow an easy and
straightforward implementation of the Decoder and Encoder traits from
tokio-util, as initially planned.